Added some clean up and new screenies. Got rid of useless stuffs. Added RAII for undocumented for the usermode component.
