--- Gizmo/banner.h
@@ -1,16 +1,0 @@
-#pragma once
-
-///
-/// Is it even hacker if it doesn't have some sort of ascii banner?? :)
-/// 
-char banner[] =
-"\n  ,--,   ,---.    ,---.          ,-.    ,-..-. .-.   .---. \n"
-".' .'    | .-.\\   | .-'  |\\    /|| |    |(||  \\| |  ( .-._)\n"
-"|  |  __ | `-'/   | `-.  |(\\  / || |    (_)|   | | (_) \\   \n"
-"\\  \\ ( _)|   (    | .-'  (_)\\/  || |    | || |\\  | _  \\ \\  \n"
-" \\  `-) )| |\\ \\   |  `--.| \\  / || `--. | || | |)|( `-'  ) \n"
-" )\\____/ |_| \\)\\  /( __.'| |\\/| ||( __.'`-'/(  (_) `----'  \n"
-"(__)         (__)(__)    '-'  '-'(_)      (__)             \n";
-
-
-/// EOF
