+++ Gizmo/types.h
@@ -5,12 +5,6 @@
 ///-------------------------------------------------------------------------------------------------
 /// Macaroo's
 ///-------------------------------------------------------------------------------------------------
-#define GET_NATIVE(_name_)                              \
-    _##_name_ f_##_name_ = (_##_name_)GetProcAddress(	\
-        GetModuleHandle(L"ntdll.dll"),                  \
-        (#_name_)                                       \
-    );
-
 #define ENCODE_CTL(Function) ( \
     ((0x8000) << 16) |                  \
     ((FILE_READ_ACCESS) << 14) |              \
@@ -25,27 +19,14 @@
 ///-------------------------------------------------------------------------------------------------
 /// Definitions
 ///-------------------------------------------------------------------------------------------------
-#define NT_SUCCESS(Status)          (((NTSTATUS)(Status)) >= 0)
-#define STATUS_SUCCESS              ((NTSTATUS)0x00000000L)
-#define STATUS_UNSUCCESSFUL         ((NTSTATUS)0xC0000001L)
-#define STATUS_INFO_LENGTH_MISMATCH ((NTSTATUS)0xC0000004L)
-
-#define FILE_ANY_ACCESS                 0
-#define FILE_SPECIAL_ACCESS    (FILE_ANY_ACCESS)
-#define FILE_READ_ACCESS          ( 0x0001 )    // file & pipe
-#define FILE_WRITE_ACCESS         ( 0x0002 )    // file & pipe
-
 #define METHOD_BUFFERED                 0
 #define METHOD_IN_DIRECT                1
 #define METHOD_OUT_DIRECT               2
 #define METHOD_NEITHER                  3
 
-constexpr const wchar_t* lpFileName = L"\\\\.\\Gremlins";
-constexpr const wchar_t* ServiceName = L"gremlins";
-
-/// 
-/// Syscall signature for ntdll. 
-/// 
+// 
+// Syscall signature for ntdll. 
+// 
 static UINT8 SyscallSig[] = {
     0x4c, 0x8b, 0xd1,	// mov r10, rcx
     0xb8				// mov eax, ??
@@ -63,250 +44,11 @@ enum IOCTL_FUNCTION : INT32
     Unhook = ENCODE_CTL(0x804)
 };
 
-
-enum SYSTEM_INFORMATION_CLASS
-{
-    SystemBasicInformation = 0,
-    SystemProcessorInformation = 1,
-    SystemPerformanceInformation = 2,
-    SystemTimeOfDayInformation = 3,
-    SystemPathInformation = 4,
-    SystemProcessInformation = 5,
-    SystemCallCountInformation = 6,
-    SystemDeviceInformation = 7,
-    SystemProcessorPerformanceInformation = 8,
-    SystemFlagsInformation = 9,
-    SystemCallTimeInformation = 10,
-    SystemModuleInformation = 11,
-    SystemLocksInformation = 12,
-    SystemStackTraceInformation = 13,
-    SystemPagedPoolInformation = 14,
-    SystemNonPagedPoolInformation = 15,
-    SystemHandleInformation = 16,
-    SystemObjectInformation = 17,
-    SystemPageFileInformation = 18,
-    SystemVdmInstemulInformation = 19,
-    SystemVdmBopInformation = 20,
-    SystemFileCacheInformation = 21,
-    SystemPoolTagInformation = 22,
-    SystemInterruptInformation = 23,
-    SystemDpcBehaviorInformation = 24,
-    SystemFullMemoryInformation = 25,
-    SystemLoadGdiDriverInformation = 26,
-    SystemUnloadGdiDriverInformation = 27,
-    SystemTimeAdjustmentInformation = 28,
-    SystemSummaryMemoryInformation = 29,
-    SystemMirrorMemoryInformation = 30,
-    SystemPerformanceTraceInformation = 31,
-    SystemObsolete0 = 32,
-    SystemExceptionInformation = 33,
-    SystemCrashDumpStateInformation = 34,
-    SystemKernelDebuggerInformation = 35,
-    SystemContextSwitchInformation = 36,
-    SystemRegistryQuotaInformation = 37,
-    SystemExtendServiceTableInformation = 38,
-    SystemPrioritySeperation = 39,
-    SystemVerifierAddDriverInformation = 40,
-    SystemVerifierRemoveDriverInformation = 41,
-    SystemProcessorIdleInformation = 42,
-    SystemLegacyDriverInformation = 43,
-    SystemCurrentTimeZoneInformation = 44,
-    SystemLookasideInformation = 45,
-    SystemTimeSlipNotification = 46,
-    SystemSessionCreate = 47,
-    SystemSessionDetach = 48,
-    SystemSessionInformation = 49,
-    SystemRangeStartInformation = 50,
-    SystemVerifierInformation = 51,
-    SystemVerifierThunkExtend = 52,
-    SystemSessionProcessInformation = 53,
-    SystemLoadGdiDriverInSystemSpace = 54,
-    SystemNumaProcessorMap = 55,
-    SystemPrefetcherInformation = 56,
-    SystemExtendedProcessInformation = 57,
-    SystemRecommendedSharedDataAlignment = 58,
-    SystemComPlusPackage = 59,
-    SystemNumaAvailableMemory = 60,
-    SystemProcessorPowerInformation = 61,
-    SystemEmulationBasicInformation = 62,
-    SystemEmulationProcessorInformation = 63,
-    SystemExtendedHandleInformation = 64,
-    SystemLostDelayedWriteInformation = 65,
-    SystemBigPoolInformation = 66,
-    SystemSessionPoolTagInformation = 67,
-    SystemSessionMappedViewInformation = 68,
-    SystemHotpatchInformation = 69,
-    SystemObjectSecurityMode = 70,
-    SystemWatchdogTimerHandler = 71,
-    SystemWatchdogTimerInformation = 72,
-    SystemLogicalProcessorInformation = 73,
-    SystemWow64SharedInformationObsolete = 74,
-    SystemRegisterFirmwareTableInformationHandler = 75,
-    SystemFirmwareTableInformation = 76,
-    SystemModuleInformationEx = 77,
-    SystemVerifierTriageInformation = 78,
-    SystemSuperfetchInformation = 79,
-    SystemMemoryListInformation = 80,
-    SystemFileCacheInformationEx = 81,
-    SystemThreadPriorityClientIdInformation = 82,
-    SystemProcessorIdleCycleTimeInformation = 83,
-    SystemVerifierCancellationInformation = 84,
-    SystemProcessorPowerInformationEx = 85,
-    SystemRefTraceInformation = 86,
-    SystemSpecialPoolInformation = 87,
-    SystemProcessIdInformation = 88,
-    SystemErrorPortInformation = 89,
-    SystemBootEnvironmentInformation = 90,
-    SystemHypervisorInformation = 91,
-    SystemVerifierInformationEx = 92,
-    SystemTimeZoneInformation = 93,
-    SystemImageFileExecutionOptionsInformation = 94,
-    SystemCoverageInformation = 95,
-    SystemPrefetchPatchInformation = 96,
-    SystemVerifierFaultsInformation = 97,
-    SystemSystemPartitionInformation = 98,
-    SystemSystemDiskInformation = 99,
-    SystemProcessorPerformanceDistribution = 100,
-    SystemNumaProximityNodeInformation = 101,
-    SystemDynamicTimeZoneInformation = 102,
-    SystemCodeIntegrityInformation = 103,
-    SystemProcessorMicrocodeUpdateInformation = 104,
-    SystemProcessorBrandString = 105,
-    SystemVirtualAddressInformation = 106,
-    SystemLogicalProcessorAndGroupInformation = 107,
-    SystemProcessorCycleTimeInformation = 108,
-    SystemStoreInformation = 109,
-    SystemRegistryAppendString = 110,
-    SystemAitSamplingValue = 111,
-    SystemVhdBootInformation = 112,
-    SystemCpuQuotaInformation = 113,
-    SystemNativeBasicInformation = 114,
-    SystemErrorPortTimeouts = 115,
-    SystemLowPriorityIoInformation = 116,
-    SystemBootEntropyInformation = 117,
-    SystemVerifierCountersInformation = 118,
-    SystemPagedPoolInformationEx = 119,
-    SystemSystemPtesInformationEx = 120,
-    SystemNodeDistanceInformation = 121,
-    SystemAcpiAuditInformation = 122,
-    SystemBasicPerformanceInformation = 123,
-    SystemQueryPerformanceCounterInformation = 124,
-    SystemSessionBigPoolInformation = 125,
-    SystemBootGraphicsInformation = 126,
-    SystemScrubPhysicalMemoryInformation = 127,
-    SystemBadPageInformation = 128,
-    SystemProcessorProfileControlArea = 129,
-    SystemCombinePhysicalMemoryInformation = 130,
-    SystemEntropyInterruptTimingInformation = 131,
-    SystemConsoleInformation = 132,
-    SystemPlatformBinaryInformation = 133,
-    SystemPolicyInformation = 134,
-    SystemHypervisorProcessorCountInformation = 135,
-    SystemDeviceDataInformation = 136,
-    SystemDeviceDataEnumerationInformation = 137,
-    SystemMemoryTopologyInformation = 138,
-    SystemMemoryChannelInformation = 139,
-    SystemBootLogoInformation = 140,
-    SystemProcessorPerformanceInformationEx = 141,
-    SystemCriticalProcessErrorLogInformation = 142,
-    SystemSecureBootPolicyInformation = 143,
-    SystemPageFileInformationEx = 144,
-    SystemSecureBootInformation = 145,
-    SystemEntropyInterruptTimingRawInformation = 146,
-    SystemPortableWorkspaceEfiLauncherInformation = 147,
-    SystemFullProcessInformation = 148,
-    SystemKernelDebuggerInformationEx = 149,
-    SystemBootMetadataInformation = 150,
-    SystemSoftRebootInformation = 151,
-    SystemElamCertificateInformation = 152,
-    SystemOfflineDumpConfigInformation = 153,
-    SystemProcessorFeaturesInformation = 154,
-    SystemRegistryReconciliationInformation = 155,
-    SystemEdidInformation = 156,
-    SystemManufacturingInformation = 157,
-    SystemEnergyEstimationConfigInformation = 158,
-    SystemHypervisorDetailInformation = 159,
-    SystemProcessorCycleStatsInformation = 160,
-    SystemVmGenerationCountInformation = 161,
-    SystemTrustedPlatformModuleInformation = 162,
-    SystemKernelDebuggerFlags = 163,
-    SystemCodeIntegrityPolicyInformation = 164,
-    SystemIsolatedUserModeInformation = 165,
-    SystemHardwareSecurityTestInterfaceResultsInformation = 166,
-    SystemSingleModuleInformation = 167,
-    SystemAllowedCpuSetsInformation = 168,
-    SystemVsmProtectionInformation = 169,
-    SystemInterruptCpuSetsInformation = 170,
-    SystemSecureBootPolicyFullInformation = 171,
-    SystemCodeIntegrityPolicyFullInformation = 172,
-    SystemAffinitizedInterruptProcessorInformation = 173,
-    SystemRootSiloInformation = 174,
-    SystemCpuSetInformation = 175,
-    SystemCpuSetTagInformation = 176,
-    SystemWin32WerStartCallout = 177,
-    SystemSecureKernelProfileInformation = 178,
-    SystemCodeIntegrityPlatformManifestInformation = 179,
-    SystemInterruptSteeringInformation = 180,
-    SystemSupportedProcessorArchitectures = 181,
-    SystemMemoryUsageInformation = 182,
-    SystemCodeIntegrityCertificateInformation = 183,
-    SystemPhysicalMemoryInformation = 184,
-    SystemControlFlowTransition = 185,
-    SystemKernelDebuggingAllowed = 186,
-    SystemActivityModerationExeState = 187,
-    SystemActivityModerationUserSettings = 188,
-    SystemCodeIntegrityPoliciesFullInformation = 189,
-    SystemCodeIntegrityUnlockInformation = 190,
-    SystemIntegrityQuotaInformation = 191,
-    SystemFlushInformation = 192,
-    SystemProcessorIdleMaskInformation = 193,
-    SystemSecureDumpEncryptionInformation = 194,
-    SystemWriteConstraintInformation = 195,
-    SystemKernelVaShadowInformation = 196,
-    SystemHypervisorSharedPageInformation = 197,
-    SystemFirmwareBootPerformanceInformation = 198,
-    SystemCodeIntegrityVerificationInformation = 199,
-    SystemFirmwarePartitionInformation = 200,
-    SystemSpeculationControlInformation = 201,
-    SystemDmaGuardPolicyInformation = 202,
-    SystemEnclaveLaunchControlInformation = 203,
-    SystemWorkloadAllowedCpuSetsInformation = 204,
-    SystemCodeIntegrityUnlockModeInformation = 205,
-    SystemLeapSecondInformation = 206,
-    SystemFlags2Information = 207,
-    SystemSecurityModelInformation = 208,
-    SystemCodeIntegritySyntheticCacheInformation = 209,
-    SystemFeatureConfigurationInformation = 210,
-    SystemFeatureConfigurationSectionInformation = 211,
-    SystemFeatureUsageSubscriptionInformation = 212,
-    SystemSecureSpeculationControlInformation = 213,
-    SystemSpacesBootInformation = 214,
-    SystemFwRamdiskInformation = 215,
-    SystemWheaIpmiHardwareInformation = 216,
-    SystemDifSetRuleClassInformation = 217,
-    SystemDifClearRuleClassInformation = 218,
-    SystemDifApplyPluginVerificationOnDriver = 219,
-    SystemDifRemovePluginVerificationOnDriver = 220,
-    SystemShadowStackInformation = 221,
-    SystemBuildVersionInformation = 222,
-    SystemPoolLimitInformation = 223,
-    SystemCodeIntegrityAddDynamicStore = 224,
-    SystemCodeIntegrityClearDynamicStores = 225,
-    SystemPoolZeroingInformation = 227,
-    MaxSystemInfoClass = 228
-};
-
 typedef struct _INPUT_BUFFER
 {
     UINT16 syscall;
     bool status;
 } INPUT_BUFFER, * PINPUT_BUFFER;
-
-typedef struct _SYSTEM_KERNEL_DEBUGGER_INFORMATION
-{
-    BOOLEAN DebuggerEnabled;
-    BOOLEAN DebuggerNotPresent;
-} SYSTEM_KERNEL_DEBUGGER_INFORMATION, * PSYSTEM_KERNEL_DEBUGGER_INFORMATION;
 
 
 /// EOF
