--- Gizmo/undocumented.h
@@ -1,12 +1,0 @@
-#pragma once
-#include "types.h"
-
-typedef NTSTATUS(NTAPI* _NtQuerySystemInformation)(
-    _In_ SYSTEM_INFORMATION_CLASS SystemInformationClass,
-    _Out_ PVOID SystemInformation,
-    _In_ ULONG SystemInformationLength,
-    _Out_opt_ PULONG ReturnLength
-    );
-
-
-/// EOF
