+++ Gremlins/AuxTypes.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#define AUX_POOLTAG_DBG	'dWxA'
+#define AUX_POOLTAG		'rGxA'
+
+
+/// EOF
